image/requirements.txt
bleach==6.2.0
fastapi==0.135.3
uvicorn[standard]==0.34.0
python-frontmatter==1.1.0
markdown==3.8
jinja2==3.1.6
python-multipart==0.0.20
cryptography==44.0.2
redis==5.2.1
pyyaml==6.0.2
holidays==0.67
python-dateutil==2.9.0
feedparser==6.0.11
pypdf==5.5.0
httpx==0.28.1
